move more helpers to utils
